+++ bridges/otelzap/core_test.go
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Copyright (c) 2016-2017 Uber Technologies, Inc.
+
 package otelzap
 
 import (
@@ -93,67 +94,4 @@ func TestConvertLevel(t *testing.T) {
 			t.Errorf("For level %v, expected %v but got %v", test.level, test.expectedSev, result)
 		}
 	}
-}
-
-// Copied from field_test.go. https://github.com/uber-go/zap/blob/b39f8b6b6a44d8371a87610be50cce58eeeaabcb/zapcore/memory_encoder_test.go
-func TestObjectEncoder(t *testing.T) {
-
-	tests := []struct {
-		desc     string
-		f        func(zapcore.ObjectEncoder)
-		expected interface{}
-	}{
-		{
-			desc:     "AddBinary",
-			f:        func(e zapcore.ObjectEncoder) { e.AddBinary("k", []byte("foo")) },
-			expected: []byte("foo"),
-		},
-		{
-			desc:     "AddByteString",
-			f:        func(e zapcore.ObjectEncoder) { e.AddByteString("k", []byte("foo")) },
-			expected: "foo",
-		},
-		{
-			desc:     "AddBool",
-			f:        func(e zapcore.ObjectEncoder) { e.AddBool("k", true) },
-			expected: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			enc := newObjectEncoder(1)
-			tt.f(enc)
-			require.Len(t, enc.kv, 1)
-			assert.Equal(t, tt.expected, value2Result((enc.kv[0].Value)), "Unexpected encoder output.")
-		})
-	}
-}
-
-func value2Result(v log.Value) any {
-	switch v.Kind() {
-	case log.KindBool:
-		return v.AsBool()
-	case log.KindFloat64:
-		return v.AsFloat64()
-	case log.KindInt64:
-		return v.AsInt64()
-	case log.KindString:
-		return v.AsString()
-	case log.KindBytes:
-		return v.AsBytes()
-	case log.KindSlice:
-		var s []any
-		for _, val := range v.AsSlice() {
-			s = append(s, value2Result(val))
-		}
-		return s
-	case log.KindMap:
-		m := make(map[string]any)
-		for _, val := range v.AsMap() {
-			m[val.Key] = value2Result(val.Value)
-		}
-		return m
-	}
-	return nil
 }
